+++ pyvela/pyvela/spnta.py
@@ -3,12 +3,10 @@ import json
 from typing import IO
 
 import numpy as np
-import astropy.units as u
 from pint.binaryconvert import convert_binary
 from pint.logging import setup as setup_log
 from pint.models import TimingModel, get_model_and_toas
 from pint.toa import TOAs
-from pint import dmu, DMconst
 
 from .ecorr import ecorr_sort
 from .model import pint_model_to_vela
@@ -196,10 +194,16 @@ class SPNTA:
     def scaled_toa_unceritainties(self, params: np.ndarray) -> np.ndarray:
         """Get the scaled TOA uncertainties (s) for a given set of parameters."""
         params = vl.read_params(self.model, params)
+
         ctoas = [vl.correct_toa(self.model, tvi, params) for tvi in self.toas]
+
         return np.sqrt(
             [
-                vl.value(vl.scaled_toa_error_sqr(tvi, ctoa))
+                vl.value(
+                    vl.scaled_toa_error_sqr(tvi, ctoa)
+                    if not self.is_wideband()
+                    else vl.scaled_toa_error_sqr(tvi.toa, ctoa.toa_correction)
+                )
                 for (tvi, ctoa) in zip(self.toas, ctoas)
             ]
         )
@@ -207,15 +211,22 @@ class SPNTA:
     def model_dm(self, params: np.ndarray) -> np.ndarray:
         """Compute the model DM (dmu) for a given set of parameters."""
         params = vl.read_params(self.model, params)
-        dms = np.zeros(len(self.toas))
-        for ii, toa in enumerate(self.toas):
-            ctoa = vl.TOACorrection()
-            for component in self.model.components:
-                if vl.isa(component, vl.DispersionComponent):
-                    dms[ii] += vl.value(vl.dispersion_slope(component, toa, ctoa, params))
-                ctoa = vl.correct_toa(component, toa, ctoa, params)
-        
-        dmu_conversion_factor = 2.41e-16 # Hz / (DMconst * dmu)
+
+        if not self.is_wideband():
+            dms = np.zeros(len(self.toas))
+            for ii, toa in enumerate(self.toas):
+                ctoa = vl.TOACorrection()
+                for component in self.model.components:
+                    if vl.isa(component, vl.DispersionComponent):
+                        dms[ii] += vl.value(
+                            vl.dispersion_slope(component, toa, ctoa, params)
+                        )
+                    ctoa = vl.correct_toa(component, toa, ctoa, params)
+        else:
+            ctoas = [vl.correct_toa(self.model, tvi, params) for tvi in self.toas]
+            dms = np.array([vl.value(ctoa.dm_correction.model_dm) for ctoa in ctoas])
+
+        dmu_conversion_factor = 2.41e-16  # Hz / (DMconst * dmu)
 
         return dms * dmu_conversion_factor
 
@@ -226,6 +237,40 @@ class SPNTA:
         model, toas = vl.load_pulsar_data(filename)
         spnta.jlsofile = filename
         spnta._setup(model, toas)
+        return spnta
+
+    @classmethod
+    def from_pint(
+        cls, model: TimingModel, toas: TOAs, cheat_prior_scale=20, custom_priors={}
+    ):
+        """Construct an `SPNTA` object from PINT `TimingModel` and `TOAs` objects"""
+        spnta = cls.__new__(cls)
+
+        setup_log(level="WARNING")
+
+        spnta.model_pint = model
+
+        # custom_priors_dict is in the "raw" format. The numbers may be
+        # in "normal" units and have to be converted into internal units.
+        if isinstance(custom_priors, dict):
+            custom_priors_dict = custom_priors
+        elif isinstance(custom_priors, str):
+            with open(custom_priors) as custom_priors_file:
+                custom_priors_dict = json.load(custom_priors_file)
+        else:
+            custom_priors_dict = json.load(custom_priors)
+
+        custom_priors = process_custom_priors(custom_priors_dict, model)
+
+        model_v, toas_v = convert_model_and_toas(
+            model,
+            toas,
+            cheat_prior_scale=cheat_prior_scale,
+            custom_priors=custom_priors,
+        )
+
+        spnta._setup(model_v, toas_v)
+
         return spnta
 
     def update_pint_model(self, samples: np.ndarray) -> TimingModel: